Fix PyTorch model float64/float32 type mismatch
PROBLEM FOUND:
- A/B routing IS working (50/50 split)
- Both models deployed and ready
- Keras model: 100% success rate
- PyTorch model: 0% success rate (all HTTP 400)
- Error: 'Input type (double) and bias type (float) should be the same'

ROOT CAUSE:
- Demo sends inputs as Python float (float64)
- Keras/TensorFlow auto-converts types
- PyTorch expects float32 - fails on float64

SOLUTIONS PROVIDED:
1. Added .astype(np.float32) to demo (doesn't work - .tolist() converts back to float64)
2. Created fix_pytorch_model_wrapper.py notebook to re-register model with float32 conversion wrapper

TO FIX:
1. Run notebooks/03_ml/fix_pytorch_model_wrapper.py in Databricks
2. Update terraform/databricks/endpoints.tf: entity_version = "2"
3. Run terraform apply
4. PyTorch model will work, dashboard will show both models

+++ notebooks/05_demo/end_to_end_demo.py
@@ -66,7 +66,7 @@ print("  - Security: endpoint shut down when not in use")
 print("  Waiting for cold start...")
 
 # Send a dummy request to wake up the endpoint
-dummy_data = np.random.rand(64, 64, 3).tolist()
+dummy_data = np.random.rand(64, 64, 3).astype(np.float32).tolist()
 warmup_payload = {"inputs": [dummy_data]}
 
 try:
@@ -143,7 +143,8 @@ for img in test_images:
     img_array = preprocess_image(img.file_path)
 
     # Call A/B endpoint
-    payload = {"inputs": [img_array.tolist()]}
+    # Convert to float32 for PyTorch compatibility
+    payload = {"inputs": [img_array.astype(np.float32).tolist()]}
 
     # Use longer timeout for first request (cold start), shorter for subsequent
     timeout = 180 if first_request else 60
